requirements.txt
crython
suds
requests
raven
PyGithub
jinja2
